Creating a message for git. Git cannot clone to a subdirectory

+++ warmup/Git.cs
@@ -8,29 +8,37 @@ namespace warmup
     {
         public static void Clone(Uri sourceLocation, TargetDir target)
         {
-            var psi = new ProcessStartInfo("git",
-                string.Format("clone {0} {1}", sourceLocation, target.FullPath));
-
-            psi.UseShellExecute = false;
-            psi.CreateNoWindow = true;
-            psi.RedirectStandardOutput = true;
-            psi.RedirectStandardError = true;
-
-            //todo: better error handling
-            string output, error = "";
-            using (var p = Process.Start(psi))
+            string[] separation_characters = new string[] { ".git" };
+            string[] piecesOfPath = sourceLocation.ToString().Split(separation_characters, StringSplitOptions.RemoveEmptyEntries);
+            if (piecesOfPath != null && piecesOfPath.Length > 0)
             {
-                output = p.StandardOutput.ReadToEnd();
-                error = p.StandardError.ReadToEnd();
-            }
+                string sourceLocationToGit = piecesOfPath[0] + ".git";
 
-            Console.WriteLine(output);
-            Console.WriteLine(error);
+                var psi = new ProcessStartInfo("git",
+                string.Format("clone {0} {1}", sourceLocationToGit, target.FullPath));
 
-            string git_directory = Path.Combine(target.FullPath, ".git");
-            if (Directory.Exists(git_directory))
-            {
-                Directory.Delete(git_directory);
+                psi.UseShellExecute = false;
+                psi.CreateNoWindow = true;
+                psi.RedirectStandardOutput = true;
+                psi.RedirectStandardError = true;
+
+                //todo: better error handling
+                string output, error = "";
+                using (var p = Process.Start(psi))
+                {
+                    output = p.StandardOutput.ReadToEnd();
+                    error = p.StandardError.ReadToEnd();
+                }
+
+                Console.WriteLine(output);
+                Console.WriteLine(error);
+
+                string git_directory = Path.Combine(target.FullPath, ".git");
+                if (Directory.Exists(git_directory))
+                {
+                    Directory.Delete(git_directory);
+                }
+                
             }
 
         }
